Moved sql and visitor from variables to this scope of Relation
Modified all testing and visitor code to reference the sql variable in
the new location. This keeps us from having to use injectInspector() and
_inspect() during runtime and is more flexible for the future.

Also simplified syntax used in the TestRelation.cfc and added a couple
of tests to make sure that a deep copy of the sql struct is done. This
was not the case, so I modified the clone() method to do a hard
StructCopy() on sql when cloning.