Added ability to directly query a datasource and do query-of-queries
Made lots of modifications to Relation's structure to enable queries to be
ran and lazy loaded. Modified Sql visitor to allow queries to be in FROM
clause. The where() method automatically splits off a new relation with a
query in the FROM if called on a relation that has already been executed.
All other modifiers will automatically clone() when this is done.

Lots of tests need to be added for Relation to confirm that everything will
always work as expected with the querying.