Created Parser.cfc to turn a SQL string into a tree of nodes
This will be used in all of the SQL options of Relation to break apart
strings and identify tables, columns, and string and number literals. In
some cases, these literals will be parameterized. In other cases, the
column names used will be correctly set up to be mapped to actual columns
by the Mapper class that is also in progress.