Added include() functionality to Relation
Most of logic is stored in Mapper, which uses mapIncludes() call to parse
an include statement and follow any associations that are encountered
between models. It then builds and calls join() on the relation to add
the specified associations to the query.

Also modified Relation to include models in sql.joins in buildModelArray()