Added countTotalRecords() function to Relation
This function takes the current relation and clones it for the purpose of
finding the total number of records to be returned. Any ordering, pagination,
or columns not needed from the original relation are dropped from the 'count'
query and a number for the number of total matching records is returned.