Added special query logic to MSSql queries with GROUP BY defined
This mimics the behavior of CFWheels. Instead of doing a straight query
against the database with paging, an initial query is performed to only
get the columns used in the GROUP BY clause. Once those records are
identified, another query is ran using a set of WHERE xxx IN (...)
conditions to find the exact records and save some time processing a lot
of complex aggregate columns.

This seems to actually run more slowly than CFWheels even though the
generated SQL is nearly identical. More work is needed.