Implemented parameterization in parser; better quote matching; For #37
When a relation has parameterization enabled, literals are automatically
wrapped in cfrel.nodes.Param elements at time of parse. The default for the
findAll() function in the cfwheels plugin is also to have parameterize=true.

Also, quote matching was improved, so that strings with escaped quotes
are matched. Escaped quotes are unescaped if the string literal is going to
be parameterized.

Built a lot of testing for general parsing behavior and parameterization.