Rewrote mapping logic to be lazy and more efficient. For #37
Mappings from from(), join(), and include() no longer occur immedately.
The operations are instead placed into a queue, and processed in order when
mappings are needed. This lazy behavior will allows CFWheels relations to be
built before the model runtime is completely initialized.

The functionality of the mapper has also been reduced. It now serves as a
device for inspecting model objects to find their table, alias, properties,
and primary keys. The actual mapping search trees are stored in the relation.
This means a mapper can be shared between many relations, so fewer objects
need to be created.

Finally, mappings no longer occur in bulk from the relation object. When the
visitor hits a column, table, or wildcard reference, it asks the relation to
perform the mapping. This removes an entire tree traversal that occurred
before the query SQL was built.