Changed default blank mapped component to "base". For #47
This fixes Railo 3.3.1 and doesn't seem to hurt ColdFusion 9. It is important
to have a blank CFC to generate for which objects() is requested, but there
is no intelligent ORM present. In this case, objects() will return an array
of "base" components with the same properties as the table columns.